Mockito added to POM, basic exception handling in service layer, basic tests being created

+++ REST Service/demo/src/main/java/com/example/demo/Services/DummyEntityService.java
@@ -2,19 +2,21 @@ package com.example.demo.Services;
 
 import com.example.demo.DummyObject.DummyEntity;
 import com.example.demo.Repositories.DummyEntityRepository;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 
 /**
  * Dummy Entity service responsible for operations performed on DummyEntity objects
  */
 @Service
+@Transactional
 public class DummyEntityService {
     @Autowired
     DummyEntityRepository repository;
@@ -26,7 +28,11 @@ public class DummyEntityService {
      */
     public void deleteEntity(Long id) {
 
-        repository.deleteById(id);
+        if (id != null) {
+            repository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("An ID argument must not be null.");
+        }
     }
 
     /**
@@ -39,7 +45,7 @@ public class DummyEntityService {
 
         Optional<DummyEntity> entity = repository.findById(id);
 
-        return entity.orElseGet(DummyEntity::new);
+        return entity.orElseThrow(() -> new EntityNotFoundException("An entity couldn't be found."));
     }
 
     /**
@@ -55,24 +61,20 @@ public class DummyEntityService {
     }
 
     /**
-     * Finds all entities in database
-     *
-     * @return List of found entities
-     */
-    public List<DummyEntity> searchEntities() {
-        return repository.findAll();
-    }
-
-    /**
      * Saves an entity to database
      *
      * @param entity An object saved in database
      */
     public void saveEntity(DummyEntity entity) {
-        repository.save(entity);
+
+        if (entity != null) {
+            repository.save(entity);
+        } else {
+            throw new EntityNotFoundException("An entity couldn't be saved, because it was null.");
+        }
     }
 
-    public Optional<DummyEntity> searchEntityById(Long id) {
+    public Optional<DummyEntity> findEntityForSaveOrUpdate(Long id) {
         return repository.findById(id);
     }
 }
